refactor : selectCouponButton을 버튼과 모달로 분리
