+++ src/components/button/selectCouponButton/SelectCouponButton.tsx
@@ -1,7 +1,5 @@
-import { Modal, useModal } from "choco-modal-component";
-import { useTotalDiscount } from "../../../hooks/useTotalDiscount";
-import { formatPriceWithZero } from "../../../utils/formatPrice";
-import { CouponContentSection } from "../../couponContentSection/CouponContentSection";
+import { useModal } from "choco-modal-component";
+import { SelectCouponModal } from "../../selectCouponModal/SelectCouponModal";
 import BaseButton from "../baseButton/baseButton";
 import { StyledSelectCouponButton } from "./SelectCouponButton.styled";
 
@@ -12,32 +10,13 @@ export const SelectCouponButton: React.FC = () => {
     openModal();
   };
 
-  const totalDiscountPrice = useTotalDiscount();
-  const formattedTotalDiscountPrice = formatPriceWithZero(totalDiscountPrice);
-  const buttonText = `총 ${formattedTotalDiscountPrice}원 할인 쿠폰 사용하기`;
-
-  const handleApplyButtonClick = () => {
-    closeModal();
-  };
-
   return (
     <>
       <BaseButton>
         <StyledSelectCouponButton onClick={handleButtonClick}>쿠폰 적용</StyledSelectCouponButton>
       </BaseButton>
 
-      <Modal
-        modalPosition="center"
-        title="쿠폰을 선택해 주세요"
-        closeButtonPosition="bottom"
-        isOpen={isOpen}
-        onClose={closeModal}
-        onConfirm={() => {}}
-        size="small"
-        buttonText={buttonText}
-      >
-        <CouponContentSection onApplyButtonClick={handleApplyButtonClick} />
-      </Modal>
+      <SelectCouponModal isOpen={isOpen} onClose={closeModal} />
     </>
   );
 };
